+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/message/processors/probation/ProbationEventProcessor.kt
@@ -8,13 +8,13 @@ import uk.gov.justice.digital.hmpps.personrecord.model.types.SourceSystemType.DE
 import uk.gov.justice.digital.hmpps.personrecord.service.EventKeys
 import uk.gov.justice.digital.hmpps.personrecord.service.TelemetryService
 import uk.gov.justice.digital.hmpps.personrecord.service.format.EncodingService
-import uk.gov.justice.digital.hmpps.personrecord.service.message.CreateUpdateService
+import uk.gov.justice.digital.hmpps.personrecord.service.message.TransactionalProcessor
 import uk.gov.justice.digital.hmpps.personrecord.service.type.TelemetryEventType.MESSAGE_RECEIVED
 
 @Component
 class ProbationEventProcessor(
   private val telemetryService: TelemetryService,
-  private val createUpdateService: CreateUpdateService,
+  private val transactionalProcessor: TransactionalProcessor,
   private val personRepository: PersonRepository,
   private val encodingService: EncodingService,
 ) {
@@ -31,7 +31,7 @@ class ProbationEventProcessor(
     encodingService.getProbationCase(crn).fold(
       onSuccess = {
         it?.let {
-          createUpdateService.processMessage(Person.from(it), eventType) {
+          transactionalProcessor.processMessage(Person.from(it), eventType) {
             personRepository.findByCrn(crn)
           }
         }
